requirements.txt
django==3.2.8
djangorestframework==3.12.4
pymongo==3.12.0
djongo==1.3.6
dnspython==2.1.0
django-cors-headers==3.10.0
requests==2.26.0
google-auth==2.3.3